+++ src/revert/table_account_social_network.sql
@@ -1,0 +1,7 @@
+-- Revert maevsi:table_account_social_network from pg
+
+BEGIN;
+
+DROP TABLE maevsi.account_social_network ;
+
+COMMIT;