+++ src/verify/table_account_private.sql
@@ -3,6 +3,7 @@
 BEGIN;
 
 SELECT id,
+       birth_date,
        created,
        email_address,
        email_address_verification,